+++ src/Service/Object/Command/GetObjectsCommand.php
@@ -29,15 +29,16 @@ class GetObjectsCommand
 
 
     /**
+     * @param ObjectType[]|null $types
      * @return ObjectDto[]
      */
-    public function getObjects(ObjectType $type, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
+    public function getObjects(?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
     {
         $objects = $this->ilias_database->fetchAll($this->ilias_database->query($this->getObjectQuery(
-            $type,
             null,
             null,
             null,
+            $types,
             $title,
             null,
             $in_trash