+++ src/Service/Object/Port/ObjectService.php
@@ -288,9 +288,10 @@ class ObjectService
 
 
     /**
+     * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenById(int $id, ?ObjectType $type = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenById(int $id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -298,7 +299,7 @@ class ObjectService
         )
             ->getChildrenById(
                 $id,
-                $type,
+                $types,
                 $title,
                 $ref_ids,
                 $in_trash
@@ -307,9 +308,10 @@ class ObjectService
 
 
     /**
+     * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByImportId(string $import_id, ?ObjectType $type = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByImportId(string $import_id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -317,7 +319,7 @@ class ObjectService
         )
             ->getChildrenByImportId(
                 $import_id,
-                $type,
+                $types,
                 $title,
                 $ref_ids,
                 $in_trash
@@ -326,9 +328,10 @@ class ObjectService
 
 
     /**
+     * @param ObjectType[]|null $types
      * @return ObjectDto[]|null
      */
-    public function getChildrenByRefId(int $ref_id, ?ObjectType $type = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
+    public function getChildrenByRefId(int $ref_id, ?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : ?array
     {
         return GetChildrenCommand::new(
             $this,
@@ -336,7 +339,7 @@ class ObjectService
         )
             ->getChildrenByRefId(
                 $ref_id,
-                $type,
+                $types,
                 $title,
                 $ref_ids,
                 $in_trash
@@ -381,15 +384,16 @@ class ObjectService
 
 
     /**
+     * @param ObjectType[]|null $types
      * @return ObjectDto[]
      */
-    public function getObjects(ObjectType $type, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
+    public function getObjects(?array $types = null, ?string $title = null, bool $ref_ids = false, ?bool $in_trash = null) : array
     {
         return GetObjectsCommand::new(
             $this->ilias_database
         )
             ->getObjects(
-                $type,
+                $types,
                 $title,
                 $ref_ids,
                 $in_trash